template/{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.0.1"
description = "Basic template to start development of Data as a Product."
requires-python = ">= 3.11"
dependencies = [
    "pyspark==3.5.0",
    "pytest==8.4.1",
    "ruff==0.12.8",
    "black==25.1.0",
    "black[jupyter]==25.1.0",
    "isort==6.0.1",
    "pyyaml==6.0.2",
    "databricks-sdk==0.64.0",
    "cruft==2.15.0",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "scripts"]

[project.scripts]
main = "src.main:main"
data_product_spec_reg  = "scripts.data_product_spec_reg:main"

[tool.ruff]
line-length = 88
target-version = "py311"
builtins = ["spark"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
